Finally Rendered single Texture

+++ OverCover/UV.cpp
@@ -1,5 +1,5 @@
 #include "UV.h"
-void UV::SetUV(GLubyte _u,GLubyte _v)
+void UV::SetUV(GLfloat _u,GLfloat _v)
 {
 	u=_u;
 	v=_v;
