+++ OverCover/UV.h
@@ -2,7 +2,7 @@
 #include<GL\glew.h>
 struct UV
 {
-	GLubyte u,v;
-	void SetUV(GLubyte,GLubyte);
+	GLfloat u,v;
+	void SetUV(GLfloat,GLfloat);
 };
